fix: use Docker's built-in mechanism to set capabilities
