+++ Dockerfile
@@ -1,6 +1,5 @@
 # We use cross-compilation because QEMU is slow.
 FROM --platform=${BUILDPLATFORM} golang:1.22.3-alpine3.18@sha256:d1a601b64de09e2fa38c95e55838961811d5ca11062a8f4230a5c434b3ae2a34 AS build
-ARG EXTRA_BUILD_TAGS
 ARG GIT_DESCRIBE
 ARG TARGETOS
 ARG TARGETARCH
@@ -13,7 +12,7 @@ COPY [".", "/src/"]
 # Compile the code.
 RUN \
   CGO_ENABLED=0 GOOS=${TARGETOS} GOARCH=${TARGETARCH} GOARM=${TARGETVARIANT#v} \
-  go build -tags "timetzdata,${EXTRA_BUILD_TAGS}" -trimpath -ldflags="-w -s -X main.Version=${GIT_DESCRIBE} -buildid=" \
+  go build -tags "timetzdata" -trimpath -ldflags="-w -s -X main.Version=${GIT_DESCRIBE} -buildid=" \
   -o /bin/ddns ./cmd/ddns
 
 # The minimal images contain only the program and the consolidated certificates.
