+++ internal/config/config_read.go
@@ -35,6 +35,9 @@ func (c *Config) ReadEnv(ppfmt pp.PP) bool {
 		return false
 	}
 
+	CheckDeprecatedLinuxID(ppfmt, "PUID", "user")
+	CheckDeprecatedLinuxID(ppfmt, "PGID", "group")
+
 	return true
 }
