--- internal/droproot/checker.go
@@ -1,1 +1,0 @@
-package droproot
